[WIKI CGI SCRIPT] * BUG: Fix the "search" link.
[INTERPRETER]
* NEW: Date() and Time() now takes the milliseconds as an extra optional 
  argument.
* BUG: Time() with one argument correctly converts it to a date.
* NEW: Date() and Time() now always check the range validity of their 
  arguments.

[GB.DB.FORM]
* NEW: DataControl does not change the background of mandatory field and 
  the font of primary key fields anymore.
* NEW: DateControl.Control is a new property that returns the underlying
  control used for editing the field data.

[GB.GUI.BASE]
* NEW: HSplit and VSplit have a new Border property that define if the 
  splitter separator is visible or not.
* NEW: HSplit and VSplit now take the Spacing property into account.


git-svn-id: svn://localhost/gambas/trunk@5871 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/gbx/gbx_date.h
@@ -32,14 +32,14 @@
 
 typedef
   struct {
-    short year;
-    short month;
-    short day;
-    short hour;
-    short min;
-    short sec;
-    short weekday;
-    short msec;
+    int year;
+    int month;
+    int day;
+    int hour;
+    int min;
+    int sec;
+    int weekday;
+    int msec;
     }
   DATE_SERIAL;
